requirements.txt
ffmpeg-python==0.2.0
librosa==0.9.1
natsort==8.1.0
#matplotlib
